- added profile and api_paths blocks;
